fix: enhance notification deletion logic with user authorization and error handling

+++ backend/controllers/notification.Controller.js
@@ -29,16 +29,25 @@ export const deleteNotifications = async (req, res) => {
     res.json({ message: "Notification deleted successfully" });
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
-    console.log("Error in deleteNotification function", error);
+    console.log("Error in deleteNotifications function", error);
   }
 };
 
 export const deleteNotification = async (req, res) => {
   try {
     const notificationId = req.params.id;
+    const userId = req.user._id;
+    const notification = await Notification.findById(notificationId);
+
+    if (!notification) {
+      return res.status(404).json({ error: "Notification not found" });
+    }
+
+    if (notification.to.toString() !== userId.toString()) {
+      return res.status(401).json({ error: "Unauthorized" });
+    }
 
     await Notification.findByIdAndDelete(notificationId);
-
     res.json({ message: "Notification deleted successfully" });
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
